nova7_backend/requirements.txt
    python-dotenv
    flask
    flask-mail
    flask-sqlalchemy
    flask-cors
    flask-jwt-extended
    stripe
    google-cloud-storage
    psycopg2-binary
    flask-wtf
